Modify the output directory

+++ reconstruction.h
@@ -5,13 +5,13 @@
 
 
 void GRP_minmod
-(int const running_info[], int const m, double const h, double const alp2,
+(double const running_info[], int const m, double const h, double const alp2,
  double const rho[], double const u[], double const p[], double const rhoI[], double const uI[], double const pI[],
  double rho_L[], double rho_R[], double u_L[], double u_R[], double p_L[], double p_R[],
  double D_rho_L[], double D_rho_R[], double D_u_L[], double D_u_R[], double D_p_L[], double D_p_R[]);
 
 void GRP_minmod0
-(int const running_info[], int const m, double const h, double const alp2,
+(double const running_info[], int const m, double const h, double const alp2,
  double const rho[], double const u[], double const p[],
  double rho_L[], double rho_R[], double u_L[], double u_R[], double p_L[], double p_R[],
  double D_rho_L[], double D_rho_R[], double D_u_L[], double D_u_R[], double D_p_L[], double D_p_R[]);
@@ -21,18 +21,18 @@ void GRP_minmod0
 
 
 void WENO_5_noD
-(int const running_info[], int const m, double const h, double const eps, double const alp2, double const gamma,
+(double const running_info[], int const m, double const h, double const eps, double const alp2, double const gamma,
  double const rho[], double const mom[], double const ene[],
  double rho_L[], double rho_R[], double u_L[], double u_R[], double p_L[], double p_R[]);
 
 void WENO_5
-(int const running_info[], int const m, double const h, double const eps, double const alp2, double const gamma,
+(double const running_info[], int const m, double const h, double const eps, double const alp2, double const gamma,
  double const rho[], double const mom[], double const ene[],
  double rho_L[], double rho_R[], double u_L[], double u_R[], double p_L[], double p_R[],
  double D_rho_L[], double D_rho_R[], double D_u_L[], double D_u_R[], double D_p_L[], double D_p_R[]);
 
 void WENO_5_limited
-(int const running_info[], int const m, double const h, double const eps, double const alp2, double const gamma,
+(double const running_info[], int const m, double const h, double const eps, double const alp2, double const gamma,
  double const rho[], double const mom[], double const ene[], double const rhoI[], double const uI[], double const pI[],
  double rho_L[], double rho_R[], double u_L[], double u_R[], double p_L[], double p_R[],
  double D_rho_L[], double D_rho_R[], double D_u_L[], double D_u_R[], double D_p_L[], double D_p_R[], int trouble[]);
@@ -50,7 +50,7 @@ void local_WENO_5_inter(double h, double Q[8]);
 
 
 void HWENO_5
-(int const running_info[], int const m, double const h, double const eps, double const alp2, double const gamma,
+(double const running_info[], int const m, double const h, double const eps, double const alp2, double const gamma,
  double const rho[], double const mom[], double const ene[],
  double const rhoI[], double const momI[], double const eneI[], double const uI[], double const pI[],
  double rho_L[], double rho_R[], double u_L[], double u_R[], double p_L[], double p_R[],
@@ -58,7 +58,7 @@ void HWENO_5
 
 
 void HWENO_5_limited
-(int const running_info[], int const m, double const h, double const eps, double const alp2, double const gamma,
+(double const running_info[], int const m, double const h, double const eps, double const alp2, double const gamma,
  double const rho[], double const mom[], double const ene[],
  double const rhoI[], double const momI[], double const eneI[], double const uI[], double const pI[],
  double rho_L[], double rho_R[], double u_L[], double u_R[], double p_L[], double p_R[],
